Add outside opening hours feature (#110)
Feature shows if food parcels are outside opening hours clearly.

Happens if someone deleted or edits a schedule which kicks some food
parels outside opening hours.

The point is for the admins to first change the schedule, and then
re-schedule each food parcel, instead of first having to re-schedule.

Fixed a lot of timezone and weekday related issues as well in this.